Fix undefined value
before this, there was a warning in the console when you type
one of the dynamic input field:

Warning: A component is changing an uncontrolled input of type
text to be controlled. Input elements should not switch from
uncontrolled to controlled (or vice versa). Decide between using
a controlled or uncontrolled input element for the lifetime of
the component.*